Adjusted power regen

+++ Assets/Scripts/GameplayScripts/ShipBase/ShipHealth.cs
@@ -9,7 +9,7 @@ public class ShipHealth : MonoBehaviour
     public int ShieldCapacity { get; protected set; }
     protected int shieldRegen;
     public int PowerCapacity { get; protected set; }
-    protected int powerRegen;
+    protected float powerRegen;
 
     protected float evasion;
